CAPSTONE2_DeepFake-main (1)/CAPSTONE2_DeepFake-main/requirements.txt
streamlit
numpy
scipy
scikit-learn
joblib
librosa
soundfile
openai-whisper
torch
torchvision
tiktoken
